+++ CHANGELOG.md
@@ -23,7 +23,7 @@
 
 ### New Features
 
-- Add the Mode feature that enables altering the editor's behavior depending on the document’s syntax (Go Settings > Mode settings pane to set).
+- Add the Mode feature that enables altering the editor’s behavior depending on the document’s syntax (Go Settings > Mode settings pane to set).
 - Add the Restore Defaults button to the File Drop settings.