+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Improvements
 
 - [non-AppStore ver.] Update Sparkle from 2.5.2 to 2.6.0.
+- [dev] Update Yams from 5.0.6 to 5.1.0.
 
 
 ### Fixes