+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 ### Fixes
 
 - [beta] Fix an issue on macOS 14 that multiple cursors in editors were not drawn when the editor becomes active again.
+- [beta] Add missing localization.