+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 ### Fixes
 
+- [beta] Fix a memory leak.
 - [beta] Fix an issue on macOS 14 that multiple cursors in editors were not drawn when the editor becomes active again.
 - [beta] Add missing localization.